refactor: altera nome da classe e diretório para adequar ao padrão de projeto
