--- .idea/misc.xml
@@ -1,17 +1,0 @@
-<project version="4">
-    <component name="ExternalStorageConfigurationManager" enabled="true" />
-    <component name="ProjectRootManager" default="true" languageLevel="JDK_17" project-jdk-name="jbr-17"
-        project-jdk-type="JavaSDK" version="2">
-        <output url="file://$PROJECT_DIR$/build/classes" />
-    </component>
-    <component name="ProjectType">
-        <option name="id" value="Android" />
-    </component>
-    <component name="VisualizationToolProject">
-        <option name="state">
-            <ProjectState>
-                <option name="scale" value="0.11123853211009174" />
-            </ProjectState>
-        </option>
-    </component>
-</project>
