--- app/src/main/java/com/example/amigodavizinhanca/AtvCadastroPF.java
@@ -1,16 +1,0 @@
-package model;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import com.example.amigodavizinhanca.R;
-
-public class CadastroPF extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_cadastro_pf);
-    }
-}
