--- app/src/main/res/layout/activity_cadastro_pf.xml
@@ -1,120 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:layout_height="match_parent">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="190dp"
-        android:background="@drawable/toolbar_header"/>
-
-
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:layout_gravity="center"
-        android:layout_marginStart="20dp"
-        android:layout_marginEnd="20dp"
-        android:orientation="vertical"
-        tools:context="model.CadastroPF">
-
-
-        <EditText
-            android:id="@+id/edtNome"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:autofillHints="text"
-            android:ems="10"
-            android:hint="@string/nome"
-
-            android:inputType="text" />
-
-        <EditText
-            android:id="@+id/edtEmail"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:ems="10"
-            android:inputType="text"
-            android:hint="@string/email"
-            tools:ignore="TextFields"
-            android:autofillHints="emailAddress" />
-
-
-        <EditText
-            android:id="@+id/edtCPF"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:autofillHints="string"
-            android:ems="10"
-            android:hint="@string/cpf"
-            android:inputType="number" />
-
-
-        <EditText
-            android:id="@+id/edtCEP"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:ems="10"
-            android:hint="@string/cep"
-            android:inputType="textPostalAddress"
-            android:autofillHints="String" />
-
-        <EditText
-            android:id="@+id/edtLogradouro"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:ems="10"
-            android:inputType="text"
-            android:hint="@string/logradouro"
-            android:autofillHints="String" />
-
-        <EditText
-            android:id="@+id/edtLocalidade"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:autofillHints="String"
-            android:ems="20"
-            android:hint="@string/localidade"
-            android:inputType="text" />
-
-        <EditText
-            android:id="@+id/edtUF"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:ems="10"
-            android:inputType="text"
-            android:hint="@string/uf"
-            android:autofillHints="String" />
-
-        <EditText
-            android:id="@+id/edtNumero"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:autofillHints="text"
-            android:ems="10"
-            android:hint="@string/numero"
-            android:inputType="number" />
-
-        <EditText
-            android:id="@+id/edtComplemento"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:ems="10"
-            android:inputType="text"
-            android:hint="@string/complemento"
-            android:autofillHints="text" />
-
-
-        <Button
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Cadastrar"
-            android:onClick="ClickSalvar"/>
-    </LinearLayout>
-</LinearLayout>
